backend/requirements.txt
feedparser
youtube-transcript-api
google-generativeai
pillow
playwright
python-dotenv
